migrations updated

+++ aspnet-core/src/Dairy.EntityFrameworkCore/Migrations/20200711160342_Add_Route_table.cs
@@ -11,7 +11,7 @@ namespace Dairy.Migrations
                 name: "Dairy_Route",
                 columns: table => new
                 {
-                    Id = table.Column<long>(nullable: false)
+                    Id = table.Column<int>(nullable: false)
                         .Annotation("SqlServer:Identity", "1, 1"),
                     CreationTime = table.Column<DateTime>(nullable: false),
                     CreatorUserId = table.Column<long>(nullable: true),
